Better hashing logic.

+++ cpp/Platform.Hashing/Platform.Hashing.Hash.h
@@ -41,7 +41,7 @@ namespace Platform::Hashing
 
     template<typename T> std::size_t HashRaw(const T &value, std::size_t size = 1)
     {
-        std::uint32_t hash = typeid(T).hash_code();
+        std::size_t hash = typeid(T).hash_code();
         Combine(hash, &value, size);
         return hash;
     }
